--- app/src/main/java/semaphor/swoop/activities/BaseActivity.java
@@ -1,54 +1,0 @@
-package semaphor.swoop.activities;
-
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-
-import semaphor.swoop.R;
-
-
-/**
- * Created by f22labs on 07/03/17.
- */
-
-public class BaseActivity extends AppCompatActivity {
-
-    @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-    }
-
-    public void initToolbar(Toolbar toolbar, boolean isBackEnabled) {
-        setSupportActionBar(toolbar);
-
-        if(isBackEnabled) {
-            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-            getSupportActionBar().setDisplayShowHomeEnabled(true);
-            getSupportActionBar().setHomeAsUpIndicator(R.drawable.ic_arrow_back);
-
-        }
-    }
-
-    public void initToolbar(Toolbar toolbar, String title, boolean isBackEnabled) {
-
-        setSupportActionBar(toolbar);
-
-        if(isBackEnabled) {
-            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-            getSupportActionBar().setDisplayShowHomeEnabled(true);
-            getSupportActionBar().setHomeAsUpIndicator(R.drawable.ic_arrow_back);
-
-        }
-
-        getSupportActionBar().setTitle(title);
-
-
-
-    }
-
-
-
-
-}
