+++ app/src/main/java/semaphor/swoop/activities/MainActivity.java
@@ -1,310 +1,76 @@
 package semaphor.swoop.activities;
 
-import android.os.Bundle;
-import android.support.design.widget.TabLayout;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.Toolbar;
-import android.view.LayoutInflater;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.FrameLayout;
-import android.widget.ImageView;
-
 import semaphor.swoop.R;
-import semaphor.swoop.fragments.BaseFragment;
-import semaphor.swoop.fragments.HomeFragment;
-import semaphor.swoop.fragments.NewsFragment;
-import semaphor.swoop.fragments.ProfileFragment;
-import semaphor.swoop.fragments.SearchFragment;
-import semaphor.swoop.fragments.ShareFragment;
-import semaphor.swoop.utils.FragmentHistory;
-import semaphor.swoop.utils.Utils;
-import semaphor.swoop.views.FragNavController;
+import semaphor.swoop.activities.LoginActivity;
 
-import butterknife.BindArray;
-import butterknife.BindView;
-import butterknife.ButterKnife;
-
-public class MainActivity extends BaseActivity implements BaseFragment.FragmentNavigation, FragNavController.TransactionListener, FragNavController.RootFragmentListener {
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+import android.view.MenuItem;
 
 
-    @BindView(R.id.content_frame)
-    FrameLayout contentFrame;
 
-    @BindView(R.id.toolbar)
-    Toolbar toolbar;
-
-    private int[] mTabIconsSelected = {
-            R.drawable.tab_home,
-            R.drawable.tab_search,
-            R.drawable.tab_share,
-            R.drawable.tab_news,
-            R.drawable.tab_profile};
-
-
-    @BindArray(R.array.tab_name)
-    String[] TABS;
-
-    @BindView(R.id.bottom_tab_layout)
-    TabLayout bottomTabLayout;
-
-    private FragNavController mNavController;
-
-    private FragmentHistory fragmentHistory;
+public class MainActivity extends AppCompatActivity {
 
     @Override
-    public void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_home);
 
 
-        ButterKnife.bind(this);
-
-
-        initToolbar();
-
-        initTab();
-
-        fragmentHistory = new FragmentHistory();
-
-
-        mNavController = FragNavController.newBuilder(savedInstanceState, getSupportFragmentManager(), R.id.content_frame)
-                .transactionListener(this)
-                .rootFragmentListener(this, TABS.length)
-                .build();
-
-
-        switchTab(0);
-
-        bottomTabLayout.addOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
-            @Override
-            public void onTabSelected(TabLayout.Tab tab) {
-
-                fragmentHistory.push(tab.getPosition());
-
-                switchTab(tab.getPosition());
-
-
-            }
-
-            @Override
-            public void onTabUnselected(TabLayout.Tab tab) {
-
-            }
-
-            @Override
-            public void onTabReselected(TabLayout.Tab tab) {
-
-                mNavController.clearStack();
-
-                switchTab(tab.getPosition());
-
-
-            }
-        });
-
-    }
-
-    private void initToolbar() {
-        setSupportActionBar(toolbar);
-
-
-    }
-
-    private void initTab() {
-        if (bottomTabLayout != null) {
-            for (int i = 0; i < TABS.length; i++) {
-                bottomTabLayout.addTab(bottomTabLayout.newTab());
-                TabLayout.Tab tab = bottomTabLayout.getTabAt(i);
-                if (tab != null)
-                    tab.setCustomView(getTabView(i));
-            }
-        }
-    }
-
-
-    private View getTabView(int position) {
-        View view = LayoutInflater.from(MainActivity.this).inflate(R.layout.tab_item_bottom, null);
-        ImageView icon = (ImageView) view.findViewById(R.id.tab_icon);
-        icon.setImageDrawable(Utils.setDrawableSelector(MainActivity.this, mTabIconsSelected[position], mTabIconsSelected[position]));
-        return view;
-    }
-
-
-    @Override
-    public void onStart() {
-        super.onStart();
+        Intent intent = new Intent(this, LoginActivity.class);
+        startActivity(intent);
     }
 
     @Override
-    public void onStop() {
-
-        super.onStop();
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.menu_main, menu);
+        return true;
     }
-
-
-    private void switchTab(int position) {
-        mNavController.switchTab(position);
-
-
-//        updateToolbarTitle(position);
-    }
-
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-    }
-
-
-    @Override
-    protected void onPause() {
-        super.onPause();
-    }
-
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        int id = item.getItemId();
 
-        switch (item.getItemId()) {
-
-
-            case android.R.id.home:
-
-
-                onBackPressed();
-                return true;
+        //noinspection SimplifiableIfStatement
+        if (id == R.id.action_settings) {
+            return true;
         }
-
 
         return super.onOptionsItemSelected(item);
-
     }
+    public void initToolbar(Toolbar toolbar, boolean isBackEnabled) {
+        setSupportActionBar(toolbar);
 
-    @Override
-    public void onBackPressed() {
-
-        if (!mNavController.isRootFragment()) {
-            mNavController.popFragment();
-        } else {
-
-            if (fragmentHistory.isEmpty()) {
-                super.onBackPressed();
-            } else {
-
-
-                if (fragmentHistory.getStackSize() > 1) {
-
-                    int position = fragmentHistory.popPrevious();
-
-                    switchTab(position);
-
-                    updateTabSelection(position);
-
-                } else {
-
-                    switchTab(0);
-
-                    updateTabSelection(0);
-
-                    fragmentHistory.emptyStack();
-                }
-            }
+        if(isBackEnabled) {
+            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+            getSupportActionBar().setDisplayShowHomeEnabled(true);
+            getSupportActionBar().setHomeAsUpIndicator(R.drawable.ic_arrow_back);
 
         }
     }
 
+    public void initToolbar(Toolbar toolbar, String title, boolean isBackEnabled) {
 
-    private void updateTabSelection(int currentTab){
+        setSupportActionBar(toolbar);
 
-        for (int i = 0; i <  TABS.length; i++) {
-            TabLayout.Tab selectedTab = bottomTabLayout.getTabAt(i);
-            if(currentTab != i) {
-                selectedTab.getCustomView().setSelected(false);
-            }else{
-                selectedTab.getCustomView().setSelected(true);
-            }
-        }
-    }
-
-    @Override
-    protected void onSaveInstanceState(Bundle outState) {
-        super.onSaveInstanceState(outState);
-        if (mNavController != null) {
-            mNavController.onSaveInstanceState(outState);
-        }
-    }
-
-    @Override
-    public void pushFragment(Fragment fragment) {
-        if (mNavController != null) {
-            mNavController.pushFragment(fragment);
-        }
-    }
-
-
-    @Override
-    public void onTabTransaction(Fragment fragment, int index) {
-        // If we have a backstack, show the back button
-        if (getSupportActionBar() != null && mNavController != null) {
-
-
-            updateToolbar();
+        if(isBackEnabled) {
+            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+            getSupportActionBar().setDisplayShowHomeEnabled(true);
+            getSupportActionBar().setHomeAsUpIndicator(R.drawable.ic_arrow_back);
 
         }
-    }
-
-    private void updateToolbar() {
-        getSupportActionBar().setDisplayHomeAsUpEnabled(!mNavController.isRootFragment());
-        getSupportActionBar().setDisplayShowHomeEnabled(!mNavController.isRootFragment());
-        getSupportActionBar().setHomeAsUpIndicator(R.drawable.ic_arrow_back);
-    }
-
-
-    @Override
-    public void onFragmentTransaction(Fragment fragment, FragNavController.TransactionType transactionType) {
-        //do fragmentty stuff. Maybe change title, I'm not going to tell you how to live your life
-        // If we have a backstack, show the back button
-        if (getSupportActionBar() != null && mNavController != null) {
-
-            updateToolbar();
-
-        }
-    }
-
-    @Override
-    public Fragment getRootFragment(int index) {
-        switch (index) {
-
-            case FragNavController.TAB1:
-                return new HomeFragment();
-            case FragNavController.TAB2:
-                return new SearchFragment();
-            case FragNavController.TAB3:
-                return new ShareFragment();
-            case FragNavController.TAB4:
-                return new NewsFragment();
-            case FragNavController.TAB5:
-                return new ProfileFragment();
-
-
-        }
-        throw new IllegalStateException("Need to send an index that we know");
-    }
-
-
-//    private void updateToolbarTitle(int position){
-//
-//
-//        getSupportActionBar().setTitle(TABS[position]);
-//
-//    }
-
-
-    public void updateToolbarTitle(String title) {
-
 
         getSupportActionBar().setTitle(title);
+
+
 
     }
 }
