+++ app/src/main/java/semaphor/swoop/database/SqliteTest.java
@@ -16,7 +16,7 @@ public class SqliteTest extends Activity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_home);
 
         UsersDatabaseHandler dbHandler = new UsersDatabaseHandler(this);
         // Deleting all users
