+++ app/src/main/java/semaphor/swoop/fragments/ProfileFragment.java
@@ -6,7 +6,7 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import semaphor.swoop.R;
-import semaphor.swoop.activities.MainActivity;
+import semaphor.swoop.activities.HomeActivity;
 
 import butterknife.ButterKnife;
 
@@ -30,7 +30,7 @@ public class ProfileFragment extends BaseFragment{
 
         ButterKnife.bind(this, view);
 
-        ( (MainActivity)getActivity()).updateToolbarTitle("Profile");
+        ( (HomeActivity)getActivity()).updateToolbarTitle("Profile");
 
 
         return view;
